Updated datatable and jquery to get the latest security fixes.
Updated to Datatable 2 which supports jQuery 3.x and comes with it's
own DateTime extension and no longer supports moment.js. Datatable has a
CDN link builder to include their extensions, it will require
overwritting the extra blocks for customization.

Popper.js for Bootstrap 5.2 and 5.3 was 2.11.6 and 2.11.8, so I went with
the latest patch.

NOTE: If we update to Bootstrap to 5.3, there are some breaking changes
that would affect django-wildewidgets. Navbar for one, no longer has the
.navbar-dark class and uses a data attribute instead.